requirements.txt
amqp==5.3.1
asgiref==3.7.2
billiard==4.2.2
celery==5.5.3
celery-types==0.23.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==4.2.13
djangorestframework==3.14.0
idna==3.11
kombu==5.5.4
packaging==25.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
requests==2.32.5
six==1.17.0
smbus2==0.5.0
sparkfun_qwiic_as7265x==2.0.0
sparkfun_qwiic_i2c==2.0.0
sqlparse==0.4.4
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
